feat: create separate page for admin panel
