+++ src/pages/Admin/Admin.module.css
@@ -1,7 +1,3 @@
-.container {
-  width: 100%;
-}
-
 .admin {
   display: flex;
   justify-content: space-between;
