+++ src/pages/Admin/Admin.tsx
@@ -1,0 +1,55 @@
+import { useEffect, useState } from 'react';
+import { Plus, X } from 'lucide-react';
+import { UsersTable } from '@/components/UsersTable';
+import { ModelForm } from '@/components/ModelForm';
+import { ModelsTable } from '@/components/ModelsTable';
+import { FullScreenLoader } from '@/components/FullScreenLoader';
+import { getModels, type Model } from '@/api/models';
+import { getUsers, type User } from '@/api/users';
+import classes from './Admin.module.css';
+
+export function Admin() {
+  const [models, setModels] = useState<Model[]>([]);
+  const [users, setUsers] = useState<User[]>([]);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+
+  const [modelFormOpened, setModelFormOpened] = useState<boolean>(false);
+
+  const switchModelFormOpened = () => setModelFormOpened(!modelFormOpened);
+
+  useEffect(() => {
+    Promise.all([getModels(), getUsers()])
+      .then(([models, users]) => {
+        setModels(models);
+        setUsers(users);
+      })
+      .catch((err: Error) => setError(err.message))
+      .finally(() => setLoading(false));
+  }, []);
+
+  if (loading) return <FullScreenLoader />;
+  if (error) return <span>Ошибка: {error}</span>;
+
+  return (
+    <div className={classes.admin}>
+      <div>
+        <h2>Пользователи</h2>
+        <UsersTable users={users} />
+      </div>
+
+      <div>
+        <div className={classes.modelsHeading}>
+          <h2>Модели</h2>
+          <button className={classes.addButton} onClick={switchModelFormOpened}>
+            {modelFormOpened ? <X /> : <Plus />}
+          </button>
+        </div>
+
+        {modelFormOpened && <ModelForm />}
+
+        <ModelsTable models={models} />
+      </div>
+    </div>
+  );
+}
