--- src/pages/Profile/index.ts
@@ -1,1 +1,0 @@
-export * from './Profile';
