+++ src/router.tsx
@@ -4,6 +4,7 @@ import { Home } from './pages/Home';
 import { Profile } from './pages/Profile';
 import { Chat } from './pages/Chat';
 import { Login } from './pages/Login';
+import { Admin } from './pages/Admin';
 
 export const router = createBrowserRouter([
   {
@@ -13,6 +14,7 @@ export const router = createBrowserRouter([
       { path: '', element: <Home /> },
       { path: '/profile', element: <Profile /> },
       { path: '/chat/:chatId', element: <Chat /> },
+      { path: '/admin', element: <Admin /> },
     ],
   },
   { path: '/login', element: <Login /> },
